# doc/api.yaml
openapi: 3.0.3
info:
  title: WASAText API
  description: |-
    Connect with your friends effortlessly using WASAText! Send and receive messages, 
    whether one-on-one or in groups, all from the convenience of your PC. 
    Enjoy seamless conversations with text or GIFs and easily stay in touch 
    through your private chats or group discussions.
  version: "1.0.0"
...
servers:
  - url: "http://localhost:3000"
    description: Development server
paths:
...
  /users/{userId}:
    parameters:
      - $ref: '#/components/parameters/userId'
    get:
      tags: ["Users"]
      summary: Get user profile
      description: Get user profile information
      operationId: getUserProfile
      security:
        - BearerAuth: []
      responses:
        '200':
          description: User profile retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: "user123"
                name: "Maria"
                photoUrl: "https://api.example.com/users/user123/photo"
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: ["Users"]
      summary: Update user name
      description: Update the current user's name
      operationId: setMyUserName
      security:
        - BearerAuth: []
      requestBody:
        description: New user name
        required: true
        content:
          application/json:
            schema:
              type: object
              description: Request body for updating user name
              properties:
                name:
                  type: string
                  description: New display name for the user
                  pattern: '^.*?$'
                  minLength: 3
                  maxLength: 16
              required:
                - name
      responses:
        '200':
          description: User name updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
              example:
                id: "user123"
                name: "Maria Updated"
                photoUrl: "https://api.example.com/users/user123/photo"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          description: Username already exists
          content:
            application/json:
              schema:
                type: object
                description: Error response object
                properties:
                  message:
                    type: string
                    description: Error message text
                    pattern: '^.+$'
                    minLength: 1
                    maxLength: 500
                required:
                  - message
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users/{userId}/photo:
    parameters:
      - $ref: '#/components/parameters/userId'
    put:
      tags: ["Users"]
      summary: Set user profile photo
      description: Upload or update the user's profile photo
      operationId: setMyPhoto
      security:
        - BearerAuth: []
      requestBody:
        description: Profile photo
        required: true
        content:
          image/png:
            schema:
              type: string
              format: binary
              description: PNG image file data
              minLength: 1
              maxLength: 10485760
          image/jpeg:
            schema:
              type: string
              format: binary
              description: JPEG image file data
              minLength: 1
              maxLength: 10485760
          image/gif:
            schema:
              type: string
              format: binary
              description: GIF image file data
              minLength: 1
              maxLength: 10485760
      responses:
        '200':
          description: Photo updated successfully
          content:
            application/json:
              schema:
                type: object
                description: Success response
                properties:
                  message:
                    type: string
                    description: Success message
                    pattern: '^.+$'
                    minLength: 1
                    maxLength: 100
                    example: "Photo updated successfully"
              example:
                message: "Photo updated successfully"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '413':
          description: Photo too large
          content:
            application/json:
              schema:
                type: object
                description: Error response object
                properties:
                  message:
                    type: string
                    description: Error message text
                    pattern: '^.+$'
                    minLength: 1
                    maxLength: 500
                required:
                  - message
        '415':
          description: Unsupported media type
          content:
            application/json:
              schema:
                type: object
                description: Error response object
                properties:
                  message:
                    type: string
                    description: Error message text
                    pattern: '^.+$'
                    minLength: 1
                    maxLength: 500
                required:
                  - message
        '500':
          $ref: '#/components/responses/InternalServerError'
    get:
      tags: ["Users"]
      summary: Get user profile photo
      description: Retrieve the user's profile photo
      operationId: getUserPhoto
      responses:
        '200':
          description: Photo retrieved successfully
          content:
            image/png:
              schema:
                type: string
                format: binary
                description: PNG image file data
                minLength: 1
                maxLength: 10485760
            image/jpeg:
              schema:
                type: string
                format: binary
                description: JPEG image file data
                minLength: 1
                maxLength: 10485760
            image/gif:
              schema:
                type: string
                format: binary
                description: GIF image file data
                minLength: 1
                maxLength: 10485760
        '404':
          description: User not found or no photo available
          content:
            application/json:
              schema:
                type: object
                description: Error response object
                properties:
                  message:
                    type: string
                    description: Error message text
                    pattern: '^.+$'
                    minLength: 1
                    maxLength: 500
                required:
                  - message
        '500':
          $ref: '#/components/responses/InternalServerError'
  /users:
    get:
      tags: ["Users"]
      summary: Search users by username
      description: Search for users by username to start conversations
      operationId: searchUsers
      security:
        - BearerAuth: []
      parameters:
        - name: q
          in: query
          required: true
          description: Username search query
          schema:
            type: string
            description: Username search query
            pattern: '^[\w\s\-\.]+$'
            minLength: 1
            maxLength: 16
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                description: Search results containing matching users
                properties:
                  users:
                    type: array
                    description: List of users matching the search query
                    minItems: 0
                    maxItems: 100
                    items:
                      $ref: '#/components/schemas/User'
                required:
                  - users
              example:
                users:
                  - id: "user123"
                    name: "Maria"
                    photoUrl: "https://api.example.com/users/user123/photo"
                  - id: "user456"
                    name: "Marco"
                    photoUrl: "https://api.example.com/users/user456/photo"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
components:
  schemas:
    User:
      type: object
      description: User profile information
      properties:
        id:
          type: string
          description: Unique user identifier
          pattern: '^[a-zA-Z0-9]+$'
          minLength: 1
          maxLength: 50
          example: "user123"
        name:
          type: string
          description: User's display name
          minLength: 3
          maxLength: 16
          pattern: '^.*?$'
          example: "Maria"
        photoUrl:
          type: string
          format: url
          description: URL to user's profile photo
          pattern: '^https?://.+$'
          minLength: 1
          maxLength: 500
          example: "https://api.example.com/users/user123/photo"
      required:
        - id
        - name
...
  responses:
    BadRequest:
      description: The request was not compliant with the documentation (e.g. missing fields, etc.)
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    Unauthorized:
      description: The access token is missing or invalid
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    Forbidden:
      description: The user is not authorized to perform this operation
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
      content:
        application/json:
          schema:
            type: object
            description: Error response object
            properties:
              message:
                type: string
                description: Error message text
                pattern: '^.+$'
                minLength: 1
                maxLength: 500
            required:
              - message
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |-
        Bearer authentication using user identifier.
        Use the identifier returned from the login endpoint.
  parameters:
    userId:
      name: userId
      in: path
      required: true
      description: User identifier
      schema:
        type: string
        pattern: '^[a-zA-Z0-9]+$'
        minLength: 1
        maxLength: 50
        example: "user123"
...